requirements.txt
cffi==1.16.0
cryptography==41.0.7
mysql-connector==2.2.9
pycparser==2.21
python-dotenv==1.0.0
